requirements.txt
## NOTE: This is not a working req file -- merely a collection for notes.
### TODO: Make this a working file

django<1.10
djangorestframework
oauth2_provider
django-cors-headers
django-oauth-toolkit==0.11.0
django-filter
newrelic
uwsgi
pysftp
boto
pyyaml

# below is a yum install
# python-devel mysql-devel
# gcc
# mysql-python
# git
# pcre-devel
# nginx
